requirements.txt
PyYAML
Flask==1.1.2